chore: implemented luthier's create. #27

+++ src/main/java/com/tone/service/impl/LuthierServiceImpl.java
@@ -1,19 +1,21 @@
 package com.tone.service.impl;
 
-import java.util.Optional;
-import java.util.Set;
-
-import com.tone.model.enumm.StatusEnum;
-import org.springframework.stereotype.Service;
-
 import com.tone.exception.BusinessException;
 import com.tone.model.LuthierEntity;
+import com.tone.model.LuthierFeatureEntity;
 import com.tone.model.LuthierSocialNetworkEntity;
+import com.tone.model.enumm.StatusEnum;
+import com.tone.repository.LuthierFeatureRepository;
 import com.tone.repository.LuthierRepository;
 import com.tone.repository.LuthierSocialNetworkRepository;
 import com.tone.service.LuthierService;
-
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+
+import java.util.Optional;
+import java.util.Set;
+
+import static com.tone.utils.ConstantsMessages.MSG_ERROR_LUTHIER_EXIST;
 
 @Slf4j
 @Service
@@ -21,31 +23,55 @@ public class LuthierServiceImpl extends BaseServiceImpl<LuthierEntity,Long> impl
 	
 	private LuthierRepository luthierRepository;	
 	private LuthierSocialNetworkRepository luthierSocialNetworkRepository;
+	private LuthierFeatureRepository luthierFeatureRepository;
 	
-	public LuthierServiceImpl(LuthierRepository luthierRepository, LuthierSocialNetworkRepository luthierSocialNetworkRepository) {
+	public LuthierServiceImpl(LuthierRepository luthierRepository, LuthierSocialNetworkRepository luthierSocialNetworkRepository,
+							  LuthierFeatureRepository luthierFeatureRepository) {
 		super(luthierRepository);
 		this.luthierRepository = luthierRepository;
 		this.luthierSocialNetworkRepository = luthierSocialNetworkRepository;
+		this.luthierFeatureRepository = luthierFeatureRepository;
 	}
 
 	@Override
-	public LuthierEntity findOptionalByName(String name) {
-		return this.luthierRepository.findOptionalByName(name).orElse(null);
-	}	
+	public LuthierEntity findOptionalByEmail(String email) {
+		return this.luthierRepository.findOptionalByEmail(email).orElse(null);
+	}
 
+	/**
+	 * Save a luthier
+	 * @param entity Luthier will be save
+	 * @return LUthier saved (with id)
+	 * @throws BusinessException
+	 */
 	@Override
 	public LuthierEntity save(LuthierEntity entity) throws BusinessException {
 
+		Optional<LuthierEntity> luthier = this.luthierRepository.findOptionalByEmail(entity.getEmail());
+
+		if(luthier.isPresent() && !luthier.get().getId().equals(entity.getId())) {
+			throw new BusinessException(MSG_ERROR_LUTHIER_EXIST);
+		}
+
 		Set<LuthierSocialNetworkEntity> listSocial = entity.getSocialNetworks();
+		Set<LuthierFeatureEntity> features = entity.getFeatures();
 		entity.setSocialNetworks(null);
+		entity.setFeatures(null);
 
 		entity = super.save(entity);
-		
+
 		if(listSocial != null && !listSocial.isEmpty()) {
 			for(LuthierSocialNetworkEntity social : listSocial) {
 				social.setLuthier(entity);
 				social = this.luthierSocialNetworkRepository.save(social);
 				entity.addSocialNetwork(social);
+			}
+		}
+
+		if(features != null && !features.isEmpty()) {
+			for(LuthierFeatureEntity feature : features) {
+				feature = this.luthierFeatureRepository.save(new LuthierFeatureEntity(entity, feature.getFeature(), feature.getValue()));
+				entity.addFeature(feature);
 			}
 		}
 		
@@ -58,9 +84,11 @@ public class LuthierServiceImpl extends BaseServiceImpl<LuthierEntity,Long> impl
 		Optional<LuthierEntity> luthier = this.findById(entity.getId());
 		
 		if(luthier.isPresent()) {
-			
-			for(LuthierSocialNetworkEntity social : luthier.get().getSocialNetworks()) {
-				this.luthierSocialNetworkRepository.delete(social);				
+
+			if(luthier.get().getSocialNetworks() != null) {
+				for (LuthierSocialNetworkEntity social : luthier.get().getSocialNetworks()) {
+					this.luthierSocialNetworkRepository.delete(social);
+				}
 			}
 		}
